cir: Expand durations to 16-bit
Currently all durations fit in signed 8-bit variable. However, this will
change in following commit with higher clock rate.

Make durations 16-bit.

+++ drivers/cir/rc6.h
@@ -21,12 +21,12 @@ enum {
 };
 
 struct rc6_ctx {
-	const int8_t *durations;
-	uint32_t      buffer;
-	uint8_t       bits;
-	uint8_t       state;
-	uint8_t       pulse;
-	int8_t        width;
+	const int16_t *durations;
+	uint32_t       buffer;
+	uint8_t        bits;
+	uint8_t        state;
+	uint8_t        pulse;
+	int8_t         width;
 };
 
 /**
